♻️ Refactor :  getBaseCalendarType 함수 - utils/date-component 폴더로 이동
