--- src/components/data-entry/DateCalendar/DateCalendar.constants.ts
@@ -1,1 +1,0 @@
-export const CALENDARS = ['day', 'month', 'year'] as const;
