--- src/components/data-entry/DateCalendar/DateCalendar.types.ts
@@ -1,3 +1,0 @@
-import { CALENDARS } from './DateCalendar.constants';
-
-export type CalendarType = (typeof CALENDARS)[number];
