+++ src/constants/date-component.ts
@@ -3,3 +3,5 @@ export const MONTH_DIGITS = ['numeric', '2-digit', 'long', 'short'] as const;
 export const CHRONOLOGICAL_ORDER = 'chronologicalOrder' as const;
 
 export const INCLUDE_DISABLED_DATE = 'includeDisabledDate' as const;
+
+export const CALENDARS = ['day', 'month', 'year'] as const;
